fix: CodeInput component updates its state when the user overrides the input
CodeInput component now updates its state when the user override the input
instead of clearing it using backspace.
This change fixes a bug that was causing to send the old code into payload
even when the modal was displaying an updated confirmation code.
CodeInput logic upgraded from uncontrolled to controlled component.